feat: implement filter products on server
реализовал фильтрацию продуктов на сервере с помощью PredicateBuilder,
добавил параметры в http запрос

+++ Laminatoria/Infrastructure/Prices.cs
@@ -2,7 +2,7 @@
 {
     public class Prices
     {
-        public double MinPrice { get; set; }
-        public double MaxPrice { get; set; }
+        public decimal MinPrice { get; set; }
+        public decimal MaxPrice { get; set; }
     }
 }
